feat: add install and helper scripts
Provided install scripts (both bash and PS) so that the dotfiles can be
easily installed on any of my systems. PS scripts are mostly there so
that some functionality is available on Windows, especially in the Git
Bash environment.

Added a new directory ~/.bash-too.d for addon scripts to be sources in
.bashrc.

Updated .bashrc with the new infrastructure.

Moved the ssh-agent helper scripts to their own repository.